optbuilder: move finishBuildLastStmt to routine.go
When routine-related optbuilder logic was moved to a new file,
`routine.go`, the `finishBuildLastStmt` function was missed. This commit
corrects that mistake.

Release note: None
